+++ qiskit_ibm_runtime/api/rest/program_job.py
@@ -17,7 +17,7 @@ from __future__ import annotations
 import json
 from typing import TYPE_CHECKING
 
-from ...utils.json import RuntimeDecoder
+from ...json import RuntimeDecoder
 from .base import RestAdapterBase
 
 if TYPE_CHECKING:
