+++ qiskit_ibm_runtime/utils/__init__.py
@@ -28,5 +28,3 @@ from .validations import (
     validate_job_tags,
     validate_rzz_pubs,
 )
-
-from .json import RuntimeEncoder, RuntimeDecoder, to_base64_string
